Refactored Bluetooth again. Still not complete, but closer.

+++ Headers/Bluetooth/BluetoothDefinitions.h
@@ -17,7 +17,7 @@ static const std::string peerAdress( Nexus7BluetoothAddress );
 //static const uint16_t NumberOfBytesPerReceiveMessage = 12;
 //static const uint16_t NumberOfBytesPerSendMessage = 4;
 
-static const uint16_t NumberOfBytesPerReceiveMessage = 4;
-static const uint16_t NumberOfBytesPerSendMessage = 4;
+static const size_t NumberOfBytesPerReceiveMessage = 4;
+static const size_t NumberOfBytesPerSendMessage = 4;
 
 #endif // BLUETOOTHDEFINITIONS_H
